# backend/internal/interface/rest/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FSD — Free Slot Detector API",
    "version": "1.0.0",
    "description": "Group Calendar Coordination System. Supports UC1–UC12.\n\n## Authentication (Dev Mode)\n\nAll protected endpoints require the `X-User-ID` header. In production this will be replaced by a session cookie from Google OAuth.\n\n```\nX-User-ID: usr-001\n```\n\nSeeded user IDs: `usr-001` through `usr-005`.\n\n## Response Envelope\n\nAll responses follow a consistent envelope:\n\n```json\n{ \"data\": <payload>, \"error\": null }\n{ \"data\": null, \"error\": \"<message>\" }\n```"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local dev server" }
  ],
...
  "components": {
    "securitySchemes": {
      "DevUserID": {
        "type": "apiKey",
        "in": "header",
        "name": "X-User-ID",
        "description": "Dev-mode auth. Pass any seeded user ID (usr-001 … usr-005)."
      }
    },
...
  },
  "paths": {
...
    "/sync/google": {
      "post": {
        "tags": ["Sync"],
        "summary": "Sync Google Calendar (UC10)",
        "description": "Pulls the user's Google Calendar events and stores them locally. Requires a valid `gmailToken` on the user record.",
        "operationId": "syncGoogle",
        "security": [{ "DevUserID": [] }],
        "responses": {
          "200": { "description": "Sync complete" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error (e.g. Google API error)" }
        }
      }
    },
    "/sync/apple": {
      "post": {
        "tags": ["Sync"],
        "summary": "Sync Apple Calendar (UC10 — stub)",
        "description": "Apple Calendar connector is not yet configured. Returns a not-implemented error.",
        "operationId": "syncApple",
        "security": [{ "DevUserID": [] }],
        "responses": {
          "200": { "description": "Sync initiated" },
          "500": { "description": "Not implemented" }
        }
      }
    },
...
  }
}